[fix]修复播放列表和收藏列表数据不同步问题

+++ entry/src/main/ets/playlist/PlaylistManager.ts
@@ -1,5 +1,10 @@
 import { Song } from '../bean/Song'
-import { LiveData, MEDIA_SESSION_CURRENT_SONG, MEDIA_SESSION_PLAYLIST } from '../extensions/LiveData'
+import {
+    LiveData,
+    MEDIA_SESSION_CURRENT_SONG,
+    MEDIA_SESSION_FAVOURITE_LIST,
+    MEDIA_SESSION_PLAYLIST
+} from '../extensions/LiveData'
 import { Logger } from '../extensions/Logger'
 import { LoopMode } from '../player/LoopMode'
 import { PlaylistDb } from './PlaylistDb'
@@ -19,6 +24,7 @@ export class PlaylistManager {
     private static sInstance: PlaylistManager = null
     private songList = new Array<Song>()
     private shuttleList = new Array<Song>()
+    private favouriteList = new Array<Song>()
     private playingIndex = 0
     private playingSong: Song = null
     private playlistDb: PlaylistDb
@@ -60,6 +66,9 @@ export class PlaylistManager {
             let songList = await this.playlistDb.getPlaylist()
             if (songList) {
                 this.songList = songList
+                this.favouriteList = this.songList.filter((song) => {
+                    return song.isFavourite == 1
+                })
                 this.onPlaylistChanged()
             }
         } catch (err) {
@@ -67,8 +76,19 @@ export class PlaylistManager {
         }
     }
 
-    updateSong(song: Song) {
+    setFavouriteState(song: Song, isFavourite: boolean) {
         this.playlistDb.update(song)
+        let i = this.indexOf(song, this.favouriteList)
+        if (isFavourite) {
+            if (i < 0) {
+                this.favouriteList.push(song)
+            }
+        } else {
+            if (i >= 0) {
+                this.favouriteList.splice(i, 1)
+            }
+        }
+        this.onPlaylistChanged()
     }
 
     addList(songs: Array<Song>) {
@@ -80,13 +100,27 @@ export class PlaylistManager {
     remove(song: Song) {
         let index = this.indexOf(song, this.songList)
         let index2 = this.indexOf(song, this.shuttleList)
+        if (song.isFavourite == 1) {
+            let index3 = this.indexOf(song, this.favouriteList)
+            this.favouriteList.splice(index3, 1)
+        }
         Logger.d(TAG, "remove " + index + ", " + index2)
-        if (index >= 0) {
+        if (index >= 0 && index2 >= 0) {
             this.songList.splice(index, 1)
             this.shuttleList.splice(index2, 1)
             this.playlistDb.removeSong(song)
         }
         this.onPlaylistChanged()
+    }
+
+    getRealSong(song: Song): Song {
+        for (let i = 0; i < this.songList.length; i++) {
+            let cur = this.songList[i]
+            if (cur.url == song.url) {
+                return cur
+            }
+        }
+        return null
     }
 
     private indexOf(song: Song, list: Array<Song>) {
@@ -112,6 +146,7 @@ export class PlaylistManager {
         } else {
             LiveData.setValue(MEDIA_SESSION_PLAYLIST, this.songList)
         }
+        LiveData.setValue(MEDIA_SESSION_FAVOURITE_LIST, this.favouriteList)
     }
 
     private updateIndex() {
@@ -201,12 +236,5 @@ export class PlaylistManager {
                 Logger.d(TAG, "get pre index= " + preIndex)
                 return this.songList[preIndex]
         }
-    }
-
-    getFavouriteList(): Array<Song> {
-        let favouriteList = this.songList.filter((song) => {
-            return song.isFavourite == 1
-        })
-        return favouriteList
     }
 }
